costs: add the /costs route
The /costs route was added in the file costs/costs_route.go. There are no unit
tests available for this route yet, as a testing methodology for routes
requiring a logged user and ingested bills.
Also the GetElasticSearchParam function was modified to accomodate for the changes
to the resulting JSON required by es.SimplifyCostDocument.