feat: Add discount codes as an option on Tagbot account creation
Discount codes have been added:
- A new tagbot_discount_code table has been added to the SQL database,
  and properly integrated with tagbot_user
- A discount_code field has been added to the user creation request
  body, which (if non-empty) is checked for being a valid existing
  discount code and changes the free trial length from 14 days to 2
  months
- A generate-discount-code task has been added, for generating a new
  random discount code with a specific description

There are also some major changes in the xo generated code as the latest
version has some changes in the generator (including a few externally
incompatible ones I fixed) that seem to make it impossible to combine
code from it with code from older versions. I also went and fixed the
bad usage of a few `created` fields while I was at it, except for
forgotten_password where it seems to be required.